requirements.txt
streamlit
pubchempy
wikipedia-api
rdkit-pypi